refactor: update datetime component for improved performance and clarity
- Refactored the formatDateString function in the RelativeDay component to use useCallback for better performance.
- Updated the import path for getUpdateInterval to point to utils.js for consistency.
- Cleaned up the useEffect dependencies to ensure proper updates and prevent unnecessary re-renders.
- Commented out unused code in GuideModalContent to streamline the user guide presentation.

These changes enhance the maintainability and performance of the datetime component and improve the overall user experience.
