+++ apps/desktop/src/renderer/src/components/ui/datetime/index.tsx
@@ -1,4 +1,4 @@
-import { getUpdateInterval } from "@follow/components/ui/datetime/index.js"
+import { getUpdateInterval } from "@follow/components/ui/datetime/utils.js"
 import {
   Tooltip,
   TooltipContent,
@@ -7,7 +7,7 @@ import {
 } from "@follow/components/ui/tooltip/index.js"
 import { stopPropagation } from "@follow/utils/dom"
 import dayjs from "dayjs"
-import { useEffect, useRef, useState } from "react"
+import { useCallback, useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 
 import { useGeneralSettingSelector } from "~/atoms/settings/general"
@@ -19,35 +19,39 @@ export const RelativeDay = ({ date }: { date: Date }) => {
   const { t } = useTranslation("common")
   const language = useGeneralSettingSelector((s) => s.language)
 
-  const formatDateString = (date: Date) => {
-    const now = new Date()
+  const formatDateString = useCallback(
+    (date: Date) => {
+      const now = new Date()
 
-    // Remove the time part for comparison
-    const nowDate = new Date(now.getFullYear(), now.getMonth(), now.getDate())
-    const inputDate = new Date(date.getFullYear(), date.getMonth(), date.getDate())
+      // Remove the time part for comparison
+      const nowDate = new Date(now.getFullYear(), now.getMonth(), now.getDate())
+      const inputDate = new Date(date.getFullYear(), date.getMonth(), date.getDate())
 
-    const diffTime = nowDate.getTime() - inputDate.getTime()
-    const diffDays = diffTime / (1000 * 3600 * 24)
+      const diffTime = nowDate.getTime() - inputDate.getTime()
+      const diffDays = diffTime / (1000 * 3600 * 24)
 
-    if (diffDays === 0) {
-      return t("time.today")
-    } else if (diffDays === 1) {
-      return t("time.yesterday")
-    } else {
-      let locale: Intl.Locale
+      if (diffDays === 0) {
+        return t("time.today")
+      } else if (diffDays === 1) {
+        return t("time.yesterday")
+      } else {
+        let locale: Intl.Locale
 
-      try {
-        locale = new Intl.Locale(language.replace("_", "-"))
-      } catch {
-        locale = new Intl.Locale("en-US")
+        try {
+          locale = new Intl.Locale(language.replace("_", "-"))
+        } catch {
+          locale = new Intl.Locale("en-US")
+        }
+        return date.toLocaleDateString(locale, {
+          weekday: "long",
+          month: "short",
+          day: "numeric",
+        })
       }
-      return date.toLocaleDateString(locale, {
-        weekday: "long",
-        month: "short",
-        day: "numeric",
-      })
-    }
-  }
+    },
+    [t, language],
+  )
+
   const timerRef = useRef<any>(null)
   const [dateString, setDateString] = useState<string>(() => formatDateString(date))
 
@@ -58,12 +62,14 @@ export const RelativeDay = ({ date }: { date: Date }) => {
       timerRef.current = setTimeout(() => {
         setDateString(formatDateString(date))
       }, updateInterval)
+    } else {
+      setDateString(formatDateString(date))
     }
 
     return () => {
-      clearTimeout(timerRef.current)
+      timerRef.current = clearTimeout(timerRef.current)
     }
-  }, [date, language])
+  }, [date, formatDateString, language])
 
   const formated = dayjs(date).format(formatTemplateStringShort)
   if (formated === dateString) {
