+++ packages/components/src/ui/datetime/index.tsx
@@ -5,6 +5,7 @@ import { useEffect, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 
 import { Tooltip, TooltipContent, TooltipPortal, TooltipTrigger } from "../tooltip"
+import { getUpdateInterval } from "./utils"
 
 const formatTemplateString = "lll"
 
@@ -17,27 +18,6 @@ const formatTime = (
     return dayjs(date).format(template)
   }
   return dayjs.duration(dayjs(date).diff(dayjs(), "minute"), "minute").humanize()
-}
-
-export const getUpdateInterval = (date: string | Date, relativeBeforeDay?: number) => {
-  if (!relativeBeforeDay) return null
-  const diffInSeconds = Math.abs(dayjs(date).diff(new Date(), "second"))
-  if (diffInSeconds <= 60) {
-    return 1000 // Update every second
-  }
-  const diffInMinutes = Math.abs(dayjs(date).diff(new Date(), "minute"))
-  if (diffInMinutes <= 60) {
-    return 60000 // Update every minute
-  }
-  const diffInHours = Math.abs(dayjs(date).diff(new Date(), "hour"))
-  if (diffInHours <= 24) {
-    return 3600000 // Update every hour
-  }
-  const diffInDays = Math.abs(dayjs(date).diff(new Date(), "day"))
-  if (diffInDays <= relativeBeforeDay) {
-    return 86400000 // Update every day
-  }
-  return null // No need to update
 }
 
 export const RelativeTime: FC<{
